pyproject.toml
[project]
name = "cube-projection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.4",
    "pygame-ce>=2.5.3",
    "ruff>=0.11.2",
]
